Refactor : 콘서트좌석 리팩토링
